+++ Ignis-API/src/main/java/com/ignis/API/repository/CityRepository.java
@@ -8,11 +8,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.ignis.API.entity.City;
 import com.ignis.API.entity.FireStation;
 
-public interface CityRepository extends JpaRepository<City, Integer> {
+public interface CityRepository extends JpaRepository<City, Long> {
 
     // Metoda do wyszukiwania miast według jednostki strażackiej
     List<City> findByFireStation(FireStation fireStation);
 
     // Metoda do wyszukiwania miasta według nazwy i jednostki strażackiej
     Optional<City> findByNameAndFireStation(String name, FireStation fireStation);
+
 }
