+++ Ignis-API/src/main/java/com/ignis/API/repository/EmailSendRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ignis.API.entity.EmailSend;
 
-public interface EmailSendRepository extends JpaRepository<EmailSend, Integer> {
+public interface EmailSendRepository extends JpaRepository<EmailSend, Long> {
 
 }
