+++ Ignis-API/src/main/java/com/ignis/API/repository/EmailStatusRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ignis.API.entity.EmailStatus;
 
-public interface EmailStatusRepository extends JpaRepository<EmailStatus, Integer> {
+public interface EmailStatusRepository extends JpaRepository<EmailStatus, Long> {
 
     Optional<EmailStatus> findByName(String name);
 }
