+++ Ignis-API/src/main/java/com/ignis/API/repository/FirefighterActionRoleRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ignis.API.entity.FirefighterActionRole;
 
-public interface FirefighterActionRoleRepository extends JpaRepository<FirefighterActionRole, Integer> {
+public interface FirefighterActionRoleRepository extends JpaRepository<FirefighterActionRole, Long> {
 }
