+++ Ignis-API/src/main/java/com/ignis/API/repository/FirefighterRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ignis.API.entity.Firefighter;
 
-public interface FirefighterRepository extends JpaRepository<Firefighter, Integer> {
+public interface FirefighterRepository extends JpaRepository<Firefighter, Long> {
 
     Optional<Firefighter> findByUserLogin(String login);
 }
