+++ Ignis-API/src/main/java/com/ignis/API/repository/GarageRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ignis.API.entity.Garage;
 
-public interface GarageRepository extends JpaRepository<Garage, Integer> {
+public interface GarageRepository extends JpaRepository<Garage, Long> {
 }
