+++ Ignis-API/src/main/java/com/ignis/API/repository/TypeFunctionRepository.java
@@ -1,11 +1,12 @@
 package com.ignis.API.repository;
 
-import com.ignis.API.entity.TypeFunction;
-import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.Optional;
 
-public interface TypeFunctionRepository extends JpaRepository<TypeFunction, Integer> {
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.ignis.API.entity.TypeFunction;
+
+public interface TypeFunctionRepository extends JpaRepository<TypeFunction, Long> {
 
     Optional<TypeFunction> findByName(String name);
 }
