+++ Ignis-API/src/main/java/com/ignis/API/repository/UserRepository.java
@@ -2,13 +2,13 @@ package com.ignis.API.repository;
 
 import java.util.Optional;
 
-import com.ignis.API.entity.User;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.ignis.API.entity.User;
+
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     boolean existsByEmail(String email);
