+++ Ignis-API/src/main/java/com/ignis/API/repository/VehicleToCardRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ignis.API.entity.VehicleToCard;
 
-public interface VehicleToCardRepository extends JpaRepository<VehicleToCard, Integer> {
+public interface VehicleToCardRepository extends JpaRepository<VehicleToCard, Long> {
 }
